refactor(node-ui): split ProjectView.tsx into project/ subdir
ProjectView.tsx was 3395 lines, making it the largest source file in
the repository. This mechanical extraction moves the sub-components,
helpers, and local types into a sibling `project/` subdirectory while
leaving the top-level ProjectView component itself behind as a thin
orchestrator.

Scope (pure code-motion, no behavior change):

- packages/node-ui/src/ui/views/project/helpers.ts (~400 lines):
  shared types, layer/colour constants, and pure utility functions
  (LAYER_CONFIG, TRUST_COLORS, entityAuthorUri, shortType,
  useLayerTriples, formatRelativeTime, etc).

- packages/node-ui/src/ui/views/project/components.tsx (~2770 lines):
  all non-entry sub-components (ProjectHeaderStrip, LayerSwitcher,
  MemoryStrip, LayerDetailView, KADetailView, SubGraphDetailView, etc)
  and the lazy-loaded RdfGraph import.

- packages/node-ui/src/ui/views/ProjectView.tsx (292 lines):
  now just the ProjectView export + its hooks wiring + top-level
  render tree; imports everything else from ./project/.

Verified via `tsc -p src/ui/tsconfig.json --noEmit` — total error count
matches the pre-split baseline (35 on main, 35 after split), i.e. no
new type errors introduced. All extracted symbols are pure moves; the
only non-mechanical change was re-adding the VIZ_* constant imports
to helpers.ts since buildLayerGraphOptions lives there now.

This sets up ProjectView for further incremental splitting (e.g.
promoting KADetailView or SubGraphDetailView to their own files)
without the noise of today's 3k-line diff.

Made-with: Cursor
